multibox/requirements.txt
glfw==2.1.0
glm==0.4.3
numpy==1.20.2
Pillow==8.2.0
PyOpenGL==3.1.5
